edited footer ,
behind footer little bit content is hided. and I tried my best to link different sections by anchor tag to about me section and education section but its not working 100%. working littel bit when we see it on half screen of laptop but nothing hapening when we see it on full screen. I need help reguarding this.